build.gradle: specify resources directory
When trying to run the app as a gradle project, the app will crash
with a 'Exception in Application start method'.

This is due to the resources directory not being specified in
build.gradle. Thus, Main.class.getResource("ui/mainwindow.fxml") in
Gui#createMainWindow(Stage, Stoppable) is unable to find the `fxml`
resource required, causing the app to crash when trying to load a
null resource.

Let's update build.gradle to specify the resources directory.